gcp-document-apis/functions/info_extractor/requirements.txt
google-cloud-documentai==2.20.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
